feat(matrix-multipleanswer): #FOR-614 refacto directives
