+++ formulaire/src/main/resources/public/ts/directives/question/index.ts
@@ -5,5 +5,5 @@ export * from './result-question-item/result-question-item.directive';
 export * from './element-item-organization';
 export * from './question-title/question-title.directive';
 export * from './question-type';
-export * from './recap-question-item';
+export * from './recap-question-item/recap-question-item.directive';
 export * from './respond-matrix/respond-matrix.directive';
