+++ formulaire/src/main/resources/public/ts/directives/question/recap-question-item/recap-question-item.directive.ts
@@ -1,49 +1,61 @@
-import {Directive, idiom, ng, template} from "entcore";
+import {Directive, idiom, ng} from "entcore";
 import {
+    Distribution,
+    DistributionStatus,
+    Form,
     FormElements,
     Question,
     QuestionChoice,
-    Response,
+    Response, ResponseFile,
     Responses, Section,
     Types
 } from "@common/models";
+import {FORMULAIRE_EMIT_EVENT} from "@common/core/enums";
 import {IScope} from "angular";
-import {RootsConst} from "../../core/constants/roots.const";
+import {RootsConst} from "../../../core/constants/roots.const";
 
-interface IPublicRecapQuestionItemProps {
+interface IRecapQuestionItemProps {
     question: Question;
     responses: Responses;
+    form: Form;
     formElements: FormElements;
+    distribution: Distribution;
     historicPosition: number[];
 }
 
 interface IViewModel {
     types: typeof Types;
+    distributionStatus: typeof DistributionStatus;
     missingResponseHtml: string;
     otherHtml: string;
 
     getHtmlDescription(description: string) : string;
     getStringResponse(): string;
     isSelectedChoice(choice: QuestionChoice, child?: Question) : boolean;
+    getResponseFileNames() : string[];
     openQuestion(): void;
     filterQuestionResponses(): Response[];
 }
 
-interface IPublicRecapQuestionItemScope extends IScope, IPublicRecapQuestionItemProps {
+interface IRecapQuestionItemScope extends IScope, IRecapQuestionItemProps {
     vm: IViewModel;
 }
 
 class Controller implements IViewModel {
     question: Question;
     responses: Responses;
+    form: Form;
     formElements: FormElements;
+    distribution: Distribution;
     historicPosition: number[];
     types: typeof Types;
+    distributionStatus: typeof DistributionStatus;
     missingResponseHtml: string;
     otherHtml: string;
 
-    constructor(private $scope: IPublicRecapQuestionItemScope, private $sce: ng.ISCEService) {
+    constructor(private $scope: IRecapQuestionItemScope, private $sce: ng.ISCEService) {
         this.types = Types;
+        this.distributionStatus = DistributionStatus;
         this.missingResponseHtml = "<em>" + idiom.translate('formulaire.public.response.missing') + "</em>";
         this.otherHtml = "<em>" + idiom.translate('formulaire.public.other') + " : </em>";
     }
@@ -83,10 +95,20 @@ class Controller implements IViewModel {
     };
 
     isSelectedChoice = (choice: QuestionChoice, child?) : boolean => {
-        let selectedChoices: number[] = this.responses.all
+        let selectedChoices: any = this.responses.all
             .filter((r: Response) => r.question_id === this.question.id || (child && r.question_id === child.id))
             .map((r: Response) => r.choice_id);
-        return (selectedChoices as any).includes(choice.id);
+        return selectedChoices.includes(choice.id);
+    };
+
+    getResponseFileNames = () : string[] => {
+        let responses: Response[] = this.responses.all.filter((r: Response) => r.question_id === this.question.id);
+        if (responses && responses.length === 1 && responses[0].files.all.length > 0) {
+            return responses[0].files.all.map((rf: ResponseFile) => rf.filename.substring(rf.filename.indexOf("_") + 1));
+        }
+        else {
+            return [this.missingResponseHtml];
+        }
     };
 
     openQuestion = () : void => {
@@ -95,9 +117,13 @@ class Controller implements IViewModel {
             let sections: Section[] = this.formElements.getSections().all.filter((s: Section) => s.id === this.question.section_id);
             formElementPosition = sections.length === 1 ? sections[0].position : null;
         }
-        this.historicPosition = this.historicPosition.slice(0, this.historicPosition.indexOf(formElementPosition) + 1);
-        sessionStorage.setItem('historicPosition', JSON.stringify(this.historicPosition));
-        template.open('main', 'containers/respond-question');
+        let newHistoric: number[] = this.historicPosition.slice(0, this.historicPosition.indexOf(formElementPosition) + 1);
+        let data: any = {
+            path: `/form/${this.form.id}/${this.distribution.id}`,
+            position: formElementPosition,
+            historicPosition: newHistoric
+        };
+        this.$scope.$emit(FORMULAIRE_EMIT_EVENT.REDIRECT, data);
     };
 
     filterQuestionResponses = () : Response[] => {
@@ -108,7 +134,7 @@ class Controller implements IViewModel {
 function directive() {
     return {
         restrict: 'E',
-        templateUrl: `${RootsConst.directive}public-recap-question-item/public-recap-question-item.html`,
+        templateUrl: `${RootsConst.directive}/question/recap-question-item/recap-question-item.html`,
         transclude: true,
         scope: {
             question: '=',
@@ -120,7 +146,7 @@ function directive() {
         bindToController: true,
         controller: ['$scope', '$sce', Controller],
         /* interaction DOM/element */
-        link: function ($scope: IPublicRecapQuestionItemScope,
+        link: function ($scope: IRecapQuestionItemScope,
                         element: ng.IAugmentedJQuery,
                         attrs: ng.IAttributes,
                         vm: IViewModel) {
@@ -128,4 +154,4 @@ function directive() {
     }
 }
 
-export const publicRecapQuestionItem: Directive = ng.directive('publicRecapQuestionItem', directive);
+export const recapQuestionItem: Directive = ng.directive('recapQuestionItem', directive);
